+++ src/main/kotlin/com/emberjs/intl/EmberIntlHbsReferenceProvider.kt
@@ -6,8 +6,8 @@ import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceProvider
 import com.intellij.util.ProcessingContext
 
-class HbsTranslationReferenceProvider : PsiReferenceProvider() {
+class EmberIntlHbsReferenceProvider : PsiReferenceProvider() {
     override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<out PsiReference> {
-        return arrayOf(HbsTranslationReference(element, TextRange(1, element.text.length - 1)))
+        return arrayOf(EmberIntlHbsReference(element, TextRange(1, element.text.length - 1)))
     }
 }
