+++ src/test/kotlin/com/emberjs/intl/EmberIntlIndexTest.kt
@@ -6,7 +6,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.entry
 import java.nio.file.Paths
 
-class EmberTranslationIndexTest : LightPlatformCodeInsightFixtureTestCase() {
+class EmberIntlIndexTest : LightPlatformCodeInsightFixtureTestCase() {
 
     fun testSimple() = doTest("foo", mapOf("en" to "bar baz", "de" to "Bar Baz"))
     fun testPlaceholder() = doTest("long-string", mapOf("en" to "Something veeeeery long with a {placeholder}"))
@@ -17,7 +17,7 @@ class EmberTranslationIndexTest : LightPlatformCodeInsightFixtureTestCase() {
     fun testQuotes4() = doTest("quote-test4", mapOf("en" to "Foo\"bar"))
 
     fun testAllKeys() {
-        val keys = EmberTranslationIndex.getTranslationKeys(myFixture.project)
+        val keys = EmberIntlIndex.getTranslationKeys(myFixture.project)
         assertThat(keys).containsOnly("foo", "long-string", "parent.child",
                 "quote-test1", "quote-test2", "quote-test3", "quote-test4")
     }
@@ -34,11 +34,11 @@ class EmberTranslationIndexTest : LightPlatformCodeInsightFixtureTestCase() {
         myFixture.copyDirectoryToProject("fixture1", "/")
 
         // Rebuild index now that the `package.json` file is copied over
-        FileBasedIndex.getInstance().requestRebuild(EmberTranslationIndex.NAME)
+        FileBasedIndex.getInstance().requestRebuild(EmberIntlIndex.NAME)
     }
 
     private fun doTest(key: String, expected: Map<String, String>) {
-        val translations = EmberTranslationIndex.getTranslations(key, myFixture.project)
+        val translations = EmberIntlIndex.getTranslations(key, myFixture.project)
         val _expected = expected.entries.map { entry(it.key, it.value) }.toTypedArray()
         assertThat(translations).containsOnly(*_expected)
     }
